Fix tags for stripe endpoints
